# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  compareAtPrice?: number | null;
  moq: number;
  rating: number;
  reviews: number;
  stock: number;
  image: string;
  supplierId: number;
  supplierName: string;
  verified: boolean;
  unit: string;
  description: string;
  shipping?: string;
  featured?: boolean;
}

export interface Supplier {
  id: number;
  name: string;
  location: string;
  rating: number;
  responseRate: number;
  yearsActive: number;
  verified: boolean;
  productCount: number;
  image: string;
  specialty: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  supplierName: string;
}

export type OrderStatusProperty = typeof OrderStatusProperty[keyof typeof OrderStatusProperty];


export const OrderStatusProperty = {
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  date: string;
  status: OrderStatusProperty;
  total: number;
  itemCount: number;
  buyerName: string;
  destination: string;
  items: OrderItem[];
}

export interface OrderInput {
  destination: string;
}

export interface SalesPoint {
  date?: string;
  label: string;
  value: number;
}

export interface SupplierDashboard {
  revenue: number;
  revenueChange: number;
  orders: number;
  ordersChange: number;
  products: number;
  lowStock: number;
  recentOrders: Order[];
  sales: SalesPoint[];
}

export interface Category {
  id: string;
  name: string;
  count: number;
  image: string;
}

export interface CartItem {
  productId: number;
  product: Product;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  shipping: number;
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity?: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface MarketplaceSummary {
  productCount: number;
  supplierCount: number;
  countries: number;
  dealsToday: number;
}

export interface SupplierProductInput {
  name: string;
  category: string;
  price: number;
  moq: number;
  stock?: number;
  description: string;
  unit: string;
  image?: string;
}

export interface SupplierProductUpdate {
  name?: string;
  price?: number;
  moq?: number;
  stock?: number;
  description?: string;
  unit?: string;
}

export type ListProductsParams = {
/**
 * Maximum number of results
 */
limit?: number;
/**
 * Number of results to skip
 */
offset?: number;
/**
 * Filter by category
 */
category?: string;
/**
 * Search query
 */
search?: string;
/**
 * Sort order
 */
sort?: ListProductsSort;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  'price-low': 'price-low',
  'price-high': 'price-high',
  rating: 'rating',
  featured: 'featured',
} as const;